+++ MaxMind.MinFraud.UnitTest/Response/EmailTest.cs
@@ -11,12 +11,19 @@ namespace MaxMind.MinFraud.UnitTest.Response
         {
             var email = new JObject
             {
+                {
+                    "domain", 
+                    new JObject {
+                        { "first_seen", "2014-02-03" }
+                    }
+                },
                 {"first_seen", "2017-01-02" },
                 {"is_disposable", true},
                 {"is_free", true},
                 {"is_high_risk", true}
             }.ToObject<Email>()!;
 
+            Assert.Equal("2014-02-03", email.Domain.FirstSeen?.ToString("yyyy-MM-dd"));
             Assert.Equal("2017-01-02", email.FirstSeen?.ToString("yyyy-MM-dd"));
             Assert.True(email.IsDisposable);
             Assert.True(email.IsFree);